+++ tests/tests_table.py
@@ -24,6 +24,7 @@ import unittest
 
 from collections import OrderedDict
 
+import mock
 import six
 
 import rows
@@ -297,25 +298,27 @@ class TableTestCase(unittest.TestCase):
         expected = '<rows.Table 2 fields, 3 rows>'
         self.assertEqual(expected, repr(self.table))
 
-    def test_table_add_time(self):
-        '''rows.Table.__add__ should be constant time
+    def test_table_add_should_not_iterate_over_rows(self):
+        table1 = rows.Table(fields={'f1': rows.fields.IntegerField,
+                                    'f2': rows.fields.FloatField})
+        table2 = rows.Table(fields={'f1': rows.fields.IntegerField,
+                                    'f2': rows.fields.FloatField})
+        table1._rows = mock.Mock()
+        table1._rows.__add__ = mock.Mock()
+        table1._rows.__iter__ = mock.Mock()
+        table2._rows = mock.Mock()
+        table2._rows.__add__ = mock.Mock()
+        table2._rows.__iter__ = mock.Mock()
 
-        To test it we double table size for each round and then compare the
-        standard deviation to the mean (it will be almost the mean if the
-        algorithm is not fast enough and almost 10% of the mean if it's good).
-        '''
-        rounds = []
-        table = utils.table
-        for _ in range(10):
-            start = time.time()
-            table = table + table
-            end = time.time()
-            rounds.append(end - start)
-
-        mean = sum(rounds) / len(rounds)
-        stdev = math.sqrt((1.0 / (len(rounds) - 1)) *
-                          sum((value - mean) ** 2 for value in rounds))
-        self.assertTrue(0.2 * mean > stdev)
+        self.assertFalse(table1._rows.__add__.called)
+        self.assertFalse(table2._rows.__add__.called)
+        self.assertFalse(table1._rows.__iter__.called)
+        self.assertFalse(table2._rows.__iter__.called)
+        table1 + table2
+        self.assertTrue(table1._rows.__add__.called)
+        self.assertFalse(table2._rows.__add__.called)
+        self.assertFalse(table1._rows.__iter__.called)
+        self.assertFalse(table2._rows.__iter__.called)
 
 
 class TestFlexibleTable(unittest.TestCase):
@@ -374,3 +377,16 @@ class TestFlexibleTable(unittest.TestCase):
         with self.assertRaises(ValueError) as context_manager:
             self.table[[1]]
         self.assertEqual(type(context_manager.exception), ValueError)
+
+    def test_table_iadd(self):
+        table = rows.Table(fields={'f1': rows.fields.IntegerField,
+                                   'f2': rows.fields.FloatField})
+        table.append({'f1': 1, 'f2': 2})
+        table.append({'f1': 3, 'f2': 4})
+
+        self.assertEqual(len(table), 2)
+        table += table
+        self.assertEqual(len(table), 4)
+        data_rows = list(table)
+        self.assertEqual(data_rows[0], data_rows[2])
+        self.assertEqual(data_rows[1], data_rows[3])